Added k and which to svd, incorporated LOBPCG
